chore: Refactoring app/utils/database directory, simplified tests
- Changed the names of files to reflect best practices
- Changed the mock static data used in testing, made simpler
